% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rename.R
\name{renameFasta}
\alias{renameFasta}
\title{Renombrar encabezados de archivo fasta.}
\usage{
renameFasta(fastafile)
}
\arguments{
\item{fastafile}{archivo fasta}
}
\value{
Archivo fasta con los encabezados modificados
}
\description{
Función para cambiar los encabezados de archvios fasta de tal forma que quede
sólo el nombre de los organismos.
}
\examples{
fastapath <- system.file("extdata", "Protseq.txt", package = "fastaformat")
renameFasta(fastapath)
}
